requirements.txt
sentence-transformers
scikit-learn
tqdm
python-dotenv 